requirements.txt
albumentations
numpy
torch>=1.0
torchvision
matplotlib
tensorflow
tensorboard
terminaltables
pillow
tqdm
opencv-python
pandas
matplotlib
seaborn
pillow
nms
scipy
scikit-learn

# Testing
tensorboardX
tb-nightly
future